Move SPI bus handling to DeviceController
